Cleanup abstract database tests
 - collapsed all the abstract database tests into parent class

This is mostly to allow further refactoring, but it's a good PR at this point.
