+++ storage/src/test/java/tech/pegasys/teku/storage/server/kvstore/AbstractDatabaseTest.java
@@ -11,11 +11,13 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.storage.server;
+package tech.pegasys.teku.storage.server.kvstore;
 
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.Mockito.mock;
 import static tech.pegasys.teku.infrastructure.async.SafeFutureAssert.assertThatSafeFuture;
 import static tech.pegasys.teku.infrastructure.unsigned.UInt64.ONE;
 import static tech.pegasys.teku.infrastructure.unsigned.UInt64.ZERO;
@@ -23,6 +25,12 @@ import static tech.pegasys.teku.spec.config.SpecConfig.GENESIS_SLOT;
 import static tech.pegasys.teku.storage.store.StoreAssertions.assertStoresMatch;
 
 import com.google.common.collect.Streams;
+import com.google.common.io.MoreFiles;
+import com.google.common.io.RecursiveDeleteOption;
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -33,19 +41,25 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.tuweni.bytes.Bytes32;
+import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
 import tech.pegasys.teku.bls.BLSKeyGenerator;
 import tech.pegasys.teku.bls.BLSKeyPair;
+import tech.pegasys.teku.dataproviders.lookup.BlockProvider;
+import tech.pegasys.teku.dataproviders.lookup.StateAndBlockSummaryProvider;
 import tech.pegasys.teku.ethereum.pow.api.MinGenesisTimeBlockEvent;
+import tech.pegasys.teku.infrastructure.async.AsyncRunner;
+import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.TestSpecFactory;
+import tech.pegasys.teku.spec.datastructures.blocks.BlockAndCheckpointEpochs;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBlockAndState;
 import tech.pegasys.teku.spec.datastructures.blocks.SlotAndBlockRoot;
@@ -60,10 +74,17 @@ import tech.pegasys.teku.spec.generator.ChainBuilder;
 import tech.pegasys.teku.spec.generator.ChainBuilder.BlockOptions;
 import tech.pegasys.teku.spec.generator.ChainProperties;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
+import tech.pegasys.teku.storage.api.OnDiskStoreData;
 import tech.pegasys.teku.storage.api.WeakSubjectivityUpdate;
 import tech.pegasys.teku.storage.client.RecentChainData;
+import tech.pegasys.teku.storage.server.Database;
+import tech.pegasys.teku.storage.server.ShuttingDownException;
+import tech.pegasys.teku.storage.server.StateStorageMode;
+import tech.pegasys.teku.storage.server.kvstore.dataaccess.KvStoreFinalizedDao;
+import tech.pegasys.teku.storage.server.kvstore.dataaccess.KvStoreHotDao;
 import tech.pegasys.teku.storage.storageSystem.StorageSystem;
 import tech.pegasys.teku.storage.store.StoreAssertions;
+import tech.pegasys.teku.storage.store.StoreBuilder;
 import tech.pegasys.teku.storage.store.StoreConfig;
 import tech.pegasys.teku.storage.store.UpdatableStore;
 import tech.pegasys.teku.storage.store.UpdatableStore.StoreTransaction;
@@ -97,7 +118,7 @@ public abstract class AbstractDatabaseTest {
   protected List<StorageSystem> storageSystems = new ArrayList<>();
 
   @BeforeEach
-  public void setup() {
+  public void setup() throws IOException {
     createStorage(StateStorageMode.ARCHIVE);
 
     genesisBlockAndState = chainBuilder.generateGenesis(genesisTime, true);
@@ -113,27 +134,27 @@ public abstract class AbstractDatabaseTest {
     for (StorageSystem storageSystem : storageSystems) {
       storageSystem.close();
     }
+    for (File tmpDirectory : tmpDirectories) {
+      MoreFiles.deleteRecursively(tmpDirectory.toPath(), RecursiveDeleteOption.ALLOW_INSECURE);
+    }
+    tmpDirectories.clear();
   }
-
-  // This method shouldn't be called outside of createStorage
-  protected abstract StorageSystem createStorageSystemInternal(
-      final StateStorageMode storageMode,
-      final StoreConfig storeConfig,
-      final boolean storeNonCanonicalBlocks);
 
   protected void restartStorage() {
     final StorageSystem storage = storageSystem.restarted(storageMode);
     setDefaultStorage(storage);
   }
 
-  protected StorageSystem createStorage(final StateStorageMode storageMode) {
+  protected StorageSystem createStorage(final StateStorageMode storageMode) throws IOException {
     return createStorage(storageMode, StoreConfig.createDefault(), false);
   }
 
+  // This method shouldn't be called outside of createStorage
   protected StorageSystem createStorage(
       final StateStorageMode storageMode,
       final StoreConfig storeConfig,
-      final boolean storeNonCanonicalBlocks) {
+      final boolean storeNonCanonicalBlocks)
+      throws IOException {
     this.storageMode = storageMode;
     storageSystem = createStorageSystemInternal(storageMode, storeConfig, storeNonCanonicalBlocks);
     setDefaultStorage(storageSystem);
@@ -149,7 +170,7 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void createMemoryStoreFromEmptyDatabase() {
+  public void createMemoryStoreFromEmptyDatabase() throws IOException {
     createStorage(StateStorageMode.ARCHIVE);
     assertThat(database.createMemoryStore()).isEmpty();
   }
@@ -534,7 +555,7 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void handleFinalizationWhenCacheLimitsExceeded() {
+  public void handleFinalizationWhenCacheLimitsExceeded() throws IOException {
     createStorage(StateStorageMode.ARCHIVE);
     initGenesis();
 
@@ -576,7 +597,8 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void shouldRecordOptimisticTransitionExecutionPayloadWhenFinalized_singleTransaction() {
+  public void shouldRecordOptimisticTransitionExecutionPayloadWhenFinalized_singleTransaction()
+      throws IOException {
     final SignedBlockAndState transitionBlock = generateChainWithFinalizableTransitionBlock();
     final List<SignedBlockAndState> newBlocks =
         chainBuilder
@@ -601,7 +623,7 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void shouldNotRecordTransitionExecutionPayloadWhenNotOptimistic() {
+  public void shouldNotRecordTransitionExecutionPayloadWhenNotOptimistic() throws IOException {
     final SignedBlockAndState transitionBlock = generateChainWithFinalizableTransitionBlock();
     final List<SignedBlockAndState> newBlocks =
         chainBuilder
@@ -630,7 +652,8 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void shouldRecordOptimisticTransitionExecutionPayloadWhenFinalized_multiTransaction() {
+  public void shouldRecordOptimisticTransitionExecutionPayloadWhenFinalized_multiTransaction()
+      throws IOException {
     final SignedBlockAndState transitionBlock = generateChainWithFinalizableTransitionBlock();
     final List<SignedBlockAndState> newBlocks =
         chainBuilder
@@ -658,7 +681,7 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void shouldPersistOptimisticTransitionExecutionPayload() {
+  public void shouldPersistOptimisticTransitionExecutionPayload() throws IOException {
     final SignedBlockAndState transitionBlock = generateChainWithFinalizableTransitionBlock();
     final List<SignedBlockAndState> newBlocks =
         chainBuilder
@@ -685,7 +708,7 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void shouldClearOptimisticTransitionExecutionPayload() {
+  public void shouldClearOptimisticTransitionExecutionPayload() throws IOException {
     // Record optimistic transition execution payload.
     final SignedBlockAndState transitionBlock = generateChainWithFinalizableTransitionBlock();
     final List<SignedBlockAndState> newBlocks =
@@ -716,7 +739,8 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void shouldNotRemoveOptimisticFinalizedExceptionPayloadWhenFinalizedNextUpdated() {
+  public void shouldNotRemoveOptimisticFinalizedExceptionPayloadWhenFinalizedNextUpdated()
+      throws IOException {
     final SignedBlockAndState transitionBlock = generateChainWithFinalizableTransitionBlock();
     final List<SignedBlockAndState> newBlocks =
         chainBuilder
@@ -757,7 +781,7 @@ public abstract class AbstractDatabaseTest {
    *
    * @return the merge transition block
    */
-  private SignedBlockAndState generateChainWithFinalizableTransitionBlock() {
+  private SignedBlockAndState generateChainWithFinalizableTransitionBlock() throws IOException {
     createStorage(StateStorageMode.PRUNE);
     initGenesis();
 
@@ -778,17 +802,17 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void shouldRecordFinalizedBlocksAndStates_pruneMode() {
+  public void shouldRecordFinalizedBlocksAndStates_pruneMode() throws IOException {
     testShouldRecordFinalizedBlocksAndStates(StateStorageMode.PRUNE, false);
   }
 
   @Test
-  public void shouldRecordFinalizedBlocksAndStates_archiveMode() {
+  public void shouldRecordFinalizedBlocksAndStates_archiveMode() throws IOException {
     testShouldRecordFinalizedBlocksAndStates(StateStorageMode.ARCHIVE, false);
   }
 
   @Test
-  public void testShouldRecordFinalizedBlocksAndStatesInBatchUpdate() {
+  public void testShouldRecordFinalizedBlocksAndStatesInBatchUpdate() throws IOException {
     testShouldRecordFinalizedBlocksAndStates(StateStorageMode.ARCHIVE, true);
   }
 
@@ -809,7 +833,7 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void getEarliestAvailableBlockSlot_withMissingFinalizedBlocks() {
+  public void getEarliestAvailableBlockSlot_withMissingFinalizedBlocks() throws IOException {
     // Set up database from an anchor point
     final UInt64 anchorEpoch = UInt64.valueOf(10);
     final SignedBlockAndState anchorBlockAndState =
@@ -864,17 +888,17 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void startupFromNonGenesisState_prune() {
+  public void startupFromNonGenesisState_prune() throws IOException {
     testStartupFromNonGenesisState(StateStorageMode.PRUNE);
   }
 
   @Test
-  public void startupFromNonGenesisState_archive() {
+  public void startupFromNonGenesisState_archive() throws IOException {
     testStartupFromNonGenesisState(StateStorageMode.ARCHIVE);
   }
 
   @Test
-  public void orphanedBlockStorageTest_withCanonicalBlocks() {
+  public void orphanedBlockStorageTest_withCanonicalBlocks() throws IOException {
     createStorage(storageMode, StoreConfig.createDefault(), true);
     final CreateForkChainResult forkChainResult = createForkChain(false);
     assertBlocksAvailable(
@@ -886,7 +910,7 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void orphanedBlockStorageTest_multiple() {
+  public void orphanedBlockStorageTest_multiple() throws IOException {
     createStorage(storageMode, StoreConfig.createDefault(), true);
     final ChainBuilder primaryChain = ChainBuilder.create(spec, VALIDATOR_KEYS);
     primaryChain.generateGenesis(genesisTime, true);
@@ -925,7 +949,7 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void orphanedBlockStorageTest_noCanonicalBlocks() {
+  public void orphanedBlockStorageTest_noCanonicalBlocks() throws IOException {
     createStorage(storageMode, StoreConfig.createDefault(), false);
     final CreateForkChainResult forkChainResult = createForkChain(false);
     assertBlocksUnavailable(
@@ -934,6 +958,565 @@ public abstract class AbstractDatabaseTest {
             .streamBlocksAndStates(4, forkChainResult.getFirstHotBlockSlot().longValue())
             .map(SignedBlockAndState::getRoot)
             .collect(Collectors.toList()));
+  }
+
+  private final List<File> tmpDirectories = new ArrayList<>();
+
+  protected abstract StorageSystem createStorageSystem(
+      final File tempDir,
+      final StateStorageMode storageMode,
+      final StoreConfig storeConfig,
+      final boolean storeNonCanonicalBlocks);
+
+  protected StorageSystem createStorageSystemInternal(
+      final StateStorageMode storageMode,
+      final StoreConfig storeConfig,
+      final boolean storeNonCanonicalBlocks)
+      throws IOException {
+    final Path tmpDir = Files.createTempDirectory("storageTest");
+    tmpDirectories.add(tmpDir.toFile());
+    return createStorageSystem(tmpDir.toFile(), storageMode, storeConfig, storeNonCanonicalBlocks);
+  }
+
+  protected StorageSystem createStorage(final File tempDir, final StateStorageMode storageMode) {
+    return createStorage(tempDir, storageMode, false);
+  }
+
+  protected StorageSystem createStorage(
+      final File tempDir,
+      final StateStorageMode storageMode,
+      final boolean storeNonCanonicalBlocks) {
+    this.storageMode = storageMode;
+    final StorageSystem storage =
+        createStorageSystem(
+            tempDir, storageMode, StoreConfig.createDefault(), storeNonCanonicalBlocks);
+    setDefaultStorage(storage);
+    return storage;
+  }
+
+  @Test
+  public void shouldRecreateGenesisStateOnRestart_archiveMode(@TempDir final Path tempDir)
+      throws Exception {
+    testShouldRecreateGenesisStateOnRestart(tempDir, StateStorageMode.ARCHIVE);
+  }
+
+  @Test
+  public void shouldRecreateGenesisStateOnRestart_pruneMode(@TempDir final Path tempDir)
+      throws Exception {
+    testShouldRecreateGenesisStateOnRestart(tempDir, StateStorageMode.PRUNE);
+  }
+
+  public void testShouldRecreateGenesisStateOnRestart(
+      final Path tempDir, final StateStorageMode storageMode) {
+    // Set up database with genesis state
+    createStorage(tempDir.toFile(), storageMode);
+    initGenesis();
+
+    // Shutdown and restart
+    restartStorage();
+
+    final UpdatableStore memoryStore = recreateStore();
+    assertStoresMatch(memoryStore, store);
+    assertThat(database.getEarliestAvailableBlockSlot()).contains(genesisBlockAndState.getSlot());
+  }
+
+  @Test
+  public void shouldRecreateStoreOnRestart_withOffEpochBoundaryFinalizedBlock_archiveMode(
+      @TempDir final Path tempDir) throws Exception {
+    testShouldRecreateStoreOnRestartWithOffEpochBoundaryFinalizedBlock(
+        tempDir, StateStorageMode.ARCHIVE);
+  }
+
+  @Test
+  public void shouldRecreateStoreOnRestart_withOffEpochBoundaryFinalizedBlock_pruneMode(
+      @TempDir final Path tempDir) throws Exception {
+    testShouldRecreateStoreOnRestartWithOffEpochBoundaryFinalizedBlock(
+        tempDir, StateStorageMode.PRUNE);
+  }
+
+  public void testShouldRecreateStoreOnRestartWithOffEpochBoundaryFinalizedBlock(
+      final Path tempDir, final StateStorageMode storageMode) throws Exception {
+    // Set up database with genesis state
+    createStorage(tempDir.toFile(), storageMode);
+    initGenesis();
+
+    // Create finalized block at slot prior to epoch boundary
+    final UInt64 finalizedEpoch = UInt64.valueOf(2);
+    final UInt64 finalizedSlot = spec.computeStartSlotAtEpoch(finalizedEpoch).minus(UInt64.ONE);
+    chainBuilder.generateBlocksUpToSlot(finalizedSlot);
+    final SignedBlockAndState finalizedBlock = chainBuilder.getBlockAndStateAtSlot(finalizedSlot);
+    final Checkpoint finalizedCheckpoint =
+        chainBuilder.getCurrentCheckpointForEpoch(finalizedEpoch);
+
+    // Add some more blocks
+    final UInt64 firstHotBlockSlot = finalizedCheckpoint.getEpochStartSlot(spec).plus(UInt64.ONE);
+    chainBuilder.generateBlockAtSlot(firstHotBlockSlot);
+    chainBuilder.generateBlocksUpToSlot(firstHotBlockSlot.plus(10));
+
+    // Save new blocks and finalized checkpoint
+    final StoreTransaction tx = recentChainData.startStoreTransaction();
+    chainBuilder.streamBlocksAndStates(1).forEach(b -> add(tx, List.of(b)));
+    justifyAndFinalizeEpoch(finalizedCheckpoint.getEpoch(), finalizedBlock, tx);
+    tx.commit().join();
+
+    // Shutdown and restart
+    restartStorage();
+
+    final UpdatableStore memoryStore = recreateStore();
+    assertStoresMatch(memoryStore, store);
+  }
+
+  @Test
+  public void shouldPersistOnDisk_pruneMode() throws Exception {
+    testShouldPersistOnDisk(StateStorageMode.PRUNE);
+  }
+
+  @Test
+  public void shouldPersistOnDisk_archiveMode() throws Exception {
+    testShouldPersistOnDisk(StateStorageMode.ARCHIVE);
+  }
+
+  @Test
+  public void shouldRecreateAnchorStoreOnRestart(@TempDir final Path tempDir) {
+    // Set up database from an anchor point
+    final UInt64 anchorEpoch = UInt64.valueOf(10);
+    final SignedBlockAndState anchorBlockAndState =
+        chainBuilder.generateBlockAtSlot(spec.computeStartSlotAtEpoch(anchorEpoch));
+    final AnchorPoint anchor =
+        AnchorPoint.create(
+            spec, new Checkpoint(anchorEpoch, anchorBlockAndState.getRoot()), anchorBlockAndState);
+    createStorage(tempDir.toFile(), StateStorageMode.PRUNE);
+    initFromAnchor(anchor);
+
+    // Shutdown and restart
+    restartStorage();
+
+    final UpdatableStore memoryStore = recreateStore();
+    assertStoresMatch(memoryStore, store);
+    assertThat(memoryStore.getInitialCheckpoint()).contains(anchor.getCheckpoint());
+    assertThat(database.getEarliestAvailableBlockSlot()).contains(anchorBlockAndState.getSlot());
+  }
+
+  @Test
+  public void shouldThrowIfClosedDatabaseIsModified_setGenesis() throws Exception {
+    database.close();
+    assertThatThrownBy(() -> database.storeInitialAnchor(genesisAnchor))
+        .isInstanceOf(ShuttingDownException.class);
+  }
+
+  @Test
+  public void shouldThrowIfClosedDatabaseIsModified_update() throws Exception {
+    database.storeInitialAnchor(genesisAnchor);
+    database.close();
+
+    final SignedBlockAndState newValue = chainBuilder.generateBlockAtSlot(1);
+    // Sanity check
+    assertThatSafeFuture(store.retrieveBlockState(newValue.getRoot()))
+        .isCompletedWithEmptyOptional();
+    final StoreTransaction transaction = recentChainData.startStoreTransaction();
+    transaction.putBlockAndState(newValue);
+
+    final SafeFuture<Void> result = transaction.commit();
+    assertThatThrownBy(result::get).hasCauseInstanceOf(ShuttingDownException.class);
+  }
+
+  @Test
+  public void createMemoryStore_priorToGenesisTime() {
+    database.storeInitialAnchor(genesisAnchor);
+
+    final Optional<OnDiskStoreData> maybeData =
+        ((KvStoreDatabase) database).createMemoryStore(() -> 0L);
+    assertThat(maybeData).isNotEmpty();
+
+    final OnDiskStoreData data = maybeData.get();
+    final UpdatableStore store =
+        StoreBuilder.create()
+            .metricsSystem(new NoOpMetricsSystem())
+            .specProvider(spec)
+            .time(data.getTime())
+            .anchor(data.getAnchor())
+            .genesisTime(data.getGenesisTime())
+            .latestFinalized(data.getLatestFinalized())
+            .finalizedOptimisticTransitionPayload(data.getFinalizedOptimisticTransitionPayload())
+            .justifiedCheckpoint(data.getJustifiedCheckpoint())
+            .bestJustifiedCheckpoint(data.getBestJustifiedCheckpoint())
+            .blockInformation(data.getBlockInformation())
+            .votes(data.getVotes())
+            .asyncRunner(mock(AsyncRunner.class))
+            .blockProvider(mock(BlockProvider.class))
+            .stateProvider(mock(StateAndBlockSummaryProvider.class))
+            .build();
+
+    assertThat(store.getTimeSeconds()).isEqualTo(genesisTime);
+  }
+
+  @Test
+  public void shouldThrowIfClosedDatabaseIsRead_createMemoryStore() throws Exception {
+    database.storeInitialAnchor(genesisAnchor);
+    database.close();
+
+    assertThatThrownBy(database::createMemoryStore).isInstanceOf(ShuttingDownException.class);
+  }
+
+  @Test
+  public void shouldThrowIfClosedDatabaseIsRead_getSlotForFinalizedBlockRoot() throws Exception {
+    database.storeInitialAnchor(genesisAnchor);
+    database.close();
+
+    assertThatThrownBy(() -> database.getSlotForFinalizedBlockRoot(Bytes32.ZERO))
+        .isInstanceOf(ShuttingDownException.class);
+  }
+
+  @Test
+  public void shouldThrowIfClosedDatabaseIsRead_getSignedBlock() throws Exception {
+    database.storeInitialAnchor(genesisAnchor);
+    database.close();
+
+    assertThatThrownBy(() -> database.getSignedBlock(genesisCheckpoint.getRoot()))
+        .isInstanceOf(ShuttingDownException.class);
+  }
+
+  @Test
+  public void shouldThrowIfClosedDatabaseIsRead_streamFinalizedBlocks() throws Exception {
+    database.storeInitialAnchor(genesisAnchor);
+    database.close();
+
+    assertThatThrownBy(() -> database.streamFinalizedBlocks(UInt64.ZERO, UInt64.ONE))
+        .isInstanceOf(ShuttingDownException.class);
+  }
+
+  @Test
+  public void shouldThrowIfClosedDatabaseIsRead_streamFinalizedBlocksShuttingDown()
+      throws Exception {
+    database.storeInitialAnchor(genesisAnchor);
+    try (final Stream<SignedBeaconBlock> stream =
+        database.streamFinalizedBlocks(UInt64.ZERO, UInt64.valueOf(1000L))) {
+      database.close();
+      assertThatThrownBy(stream::findAny).isInstanceOf(ShuttingDownException.class);
+    }
+  }
+
+  @Test
+  public void shouldThrowIfTransactionModifiedAfterDatabaseIsClosed_updateHotDao()
+      throws Exception {
+    database.storeInitialAnchor(genesisAnchor);
+
+    try (final KvStoreHotDao.HotUpdater updater = hotUpdater()) {
+      SignedBlockAndState newBlock = chainBuilder.generateNextBlock();
+      database.close();
+      assertThatThrownBy(
+              () -> updater.addHotBlock(BlockAndCheckpointEpochs.fromBlockAndState(newBlock)))
+          .isInstanceOf(ShuttingDownException.class);
+    }
+  }
+
+  private KvStoreHotDao.HotUpdater hotUpdater() {
+    return ((KvStoreDatabase) database).dao.hotUpdater();
+  }
+
+  @Test
+  public void shouldThrowIfTransactionModifiedAfterDatabaseIsClosed_updateFinalizedDao()
+      throws Exception {
+    database.storeInitialAnchor(genesisAnchor);
+
+    try (final KvStoreFinalizedDao.FinalizedUpdater updater = finalizedUpdater()) {
+      SignedBlockAndState newBlock = chainBuilder.generateNextBlock();
+      database.close();
+      assertThatThrownBy(() -> updater.addFinalizedBlock(newBlock.getBlock()))
+          .isInstanceOf(ShuttingDownException.class);
+    }
+  }
+
+  private KvStoreFinalizedDao.FinalizedUpdater finalizedUpdater() {
+    return ((KvStoreDatabase) database).dao.finalizedUpdater();
+  }
+
+  @Test
+  public void shouldThrowIfTransactionModifiedAfterDatabaseIsClosed_updateEth1Dao()
+      throws Exception {
+    database.storeInitialAnchor(genesisAnchor);
+
+    final DataStructureUtil dataStructureUtil =
+        new DataStructureUtil(TestSpecFactory.createDefault());
+    try (final KvStoreHotDao.HotUpdater updater = hotUpdater()) {
+      final MinGenesisTimeBlockEvent genesisTimeBlockEvent =
+          dataStructureUtil.randomMinGenesisTimeBlockEvent(1);
+      database.close();
+      assertThatThrownBy(() -> updater.addMinGenesisTimeBlock(genesisTimeBlockEvent))
+          .isInstanceOf(ShuttingDownException.class);
+    }
+  }
+
+  @Test
+  public void shouldThrowIfClosedDatabaseIsRead_getHistoricalState() throws Exception {
+    // Store genesis
+    database.storeInitialAnchor(genesisAnchor);
+    // Add a new finalized block to supersede genesis
+    final SignedBlockAndState newBlock = chainBuilder.generateBlockAtSlot(1);
+    final Checkpoint newCheckpoint = getCheckpointForBlock(newBlock.getBlock());
+    final StoreTransaction transaction = recentChainData.startStoreTransaction();
+    transaction.putBlockAndState(newBlock);
+    transaction.setFinalizedCheckpoint(newCheckpoint, false);
+    transaction.commit().ifExceptionGetsHereRaiseABug();
+    // Close db
+    database.close();
+
+    assertThatThrownBy(
+            () ->
+                database.getLatestAvailableFinalizedState(
+                    genesisCheckpoint.getEpochStartSlot(spec)))
+        .isInstanceOf(ShuttingDownException.class);
+  }
+
+  @Test
+  public void shouldThrowIfTransactionModifiedAfterDatabaseIsClosedFromAnotherThread()
+      throws Exception {
+
+    for (int i = 0; i < 20; i++) {
+      createStorage(StateStorageMode.PRUNE);
+      database.storeInitialAnchor(genesisAnchor);
+
+      try (final KvStoreHotDao.HotUpdater updater = hotUpdater()) {
+        SignedBlockAndState newBlock = chainBuilder.generateNextBlock();
+
+        final Thread dbCloserThread =
+            new Thread(
+                () -> {
+                  try {
+                    database.close();
+                  } catch (Exception e) {
+                    throw new RuntimeException(e);
+                  }
+                });
+
+        dbCloserThread.start();
+        try {
+          updater.addHotBlock(BlockAndCheckpointEpochs.fromBlockAndState(newBlock));
+        } catch (Exception e) {
+          assertThat(e).isInstanceOf(ShuttingDownException.class);
+        }
+
+        dbCloserThread.join(500);
+      }
+    }
+  }
+
+  @Test
+  public void shouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart__archive(
+      @TempDir final Path tempDir) {
+    testShouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart(tempDir, StateStorageMode.ARCHIVE);
+  }
+
+  @Test
+  public void shouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart__prune(
+      @TempDir final Path tempDir) {
+    testShouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart(tempDir, StateStorageMode.PRUNE);
+  }
+
+  @Test
+  public void shouldPersistHotStates_everyEpoch() throws IOException {
+    final int storageFrequency = 1;
+    StoreConfig storeConfig =
+        StoreConfig.builder().hotStatePersistenceFrequencyInEpochs(storageFrequency).build();
+    createStorage(StateStorageMode.ARCHIVE, storeConfig, false);
+    initGenesis();
+
+    final UInt64 latestEpoch = UInt64.valueOf(3);
+    final UInt64 targetSlot = spec.computeStartSlotAtEpoch(latestEpoch);
+    chainBuilder.generateBlocksUpToSlot(targetSlot);
+
+    // Add blocks
+    addBlocks(chainBuilder.streamBlocksAndStates().collect(toList()));
+
+    // We should only be able to pull states at epoch boundaries
+    final Set<UInt64> epochBoundarySlots = getEpochBoundarySlots(1, latestEpoch.intValue());
+    for (int i = 0; i <= targetSlot.intValue(); i++) {
+      final SignedBlockAndState blockAndState = chainBuilder.getBlockAndStateAtSlot(i);
+      final Optional<BeaconState> actual = database.getHotState(blockAndState.getRoot());
+
+      if (epochBoundarySlots.contains(UInt64.valueOf(i))) {
+        assertThat(actual).contains(blockAndState.getState());
+      } else {
+        assertThat(actual).isEmpty();
+      }
+    }
+  }
+
+  @Test
+  public void shouldPersistHotStates_never() throws IOException {
+    final int storageFrequency = 0;
+    StoreConfig storeConfig =
+        StoreConfig.builder().hotStatePersistenceFrequencyInEpochs(storageFrequency).build();
+    createStorage(StateStorageMode.ARCHIVE, storeConfig, false);
+    initGenesis();
+
+    final UInt64 latestEpoch = UInt64.valueOf(3);
+    final UInt64 targetSlot = spec.computeStartSlotAtEpoch(latestEpoch);
+    chainBuilder.generateBlocksUpToSlot(targetSlot);
+
+    // Add blocks
+    addBlocks(chainBuilder.streamBlocksAndStates().collect(toList()));
+
+    for (int i = 0; i <= targetSlot.intValue(); i++) {
+      final SignedBlockAndState blockAndState = chainBuilder.getBlockAndStateAtSlot(i);
+      final Optional<BeaconState> actual = database.getHotState(blockAndState.getRoot());
+      assertThat(actual).isEmpty();
+    }
+  }
+
+  @Test
+  public void shouldPersistHotStates_everyThirdEpoch() throws IOException {
+    final int storageFrequency = 3;
+    StoreConfig storeConfig =
+        StoreConfig.builder().hotStatePersistenceFrequencyInEpochs(storageFrequency).build();
+    createStorage(StateStorageMode.ARCHIVE, storeConfig, false);
+    initGenesis();
+
+    final UInt64 latestEpoch = UInt64.valueOf(3 * storageFrequency);
+    final UInt64 targetSlot = spec.computeStartSlotAtEpoch(latestEpoch);
+    chainBuilder.generateBlocksUpToSlot(targetSlot);
+
+    // Add blocks
+    addBlocks(chainBuilder.streamBlocksAndStates().collect(toList()));
+
+    // We should only be able to pull states at epoch boundaries
+    final Set<UInt64> epochBoundarySlots = getEpochBoundarySlots(1, latestEpoch.intValue());
+    for (int i = 0; i <= targetSlot.intValue(); i++) {
+      final SignedBlockAndState blockAndState = chainBuilder.getBlockAndStateAtSlot(i);
+      final Optional<BeaconState> actual = database.getHotState(blockAndState.getRoot());
+
+      final UInt64 currentSlot = UInt64.valueOf(i);
+      final UInt64 currentEpoch = spec.computeEpochAtSlot(currentSlot);
+      final boolean shouldPersistThisEpoch = currentEpoch.mod(storageFrequency).equals(UInt64.ZERO);
+      if (epochBoundarySlots.contains(currentSlot) && shouldPersistThisEpoch) {
+        assertThat(actual).contains(blockAndState.getState());
+      } else {
+        assertThat(actual).isEmpty();
+      }
+    }
+  }
+
+  @Test
+  public void shouldClearStaleHotStates() throws IOException {
+    final int storageFrequency = 1;
+    StoreConfig storeConfig =
+        StoreConfig.builder().hotStatePersistenceFrequencyInEpochs(storageFrequency).build();
+    createStorage(StateStorageMode.ARCHIVE, storeConfig, false);
+    initGenesis();
+
+    final UInt64 latestEpoch = UInt64.valueOf(3);
+    final UInt64 targetSlot = spec.computeStartSlotAtEpoch(latestEpoch);
+    chainBuilder.generateBlocksUpToSlot(targetSlot);
+
+    // Add blocks
+    addBlocks(chainBuilder.streamBlocksAndStates().collect(toList()));
+    justifyAndFinalizeEpoch(latestEpoch, chainBuilder.getLatestBlockAndState());
+
+    // Hot states should be cleared out
+    for (int i = 0; i <= targetSlot.intValue(); i++) {
+      final SignedBlockAndState blockAndState = chainBuilder.getBlockAndStateAtSlot(i);
+      final Optional<BeaconState> actual = database.getHotState(blockAndState.getRoot());
+      assertThat(actual).isEmpty();
+    }
+  }
+
+  @Test
+  public void shouldHandleRestartWithUnrecoverableForkBlocks_archive(@TempDir final Path tempDir) {
+    testShouldHandleRestartWithUnrecoverableForkBlocks(tempDir, StateStorageMode.ARCHIVE);
+  }
+
+  @Test
+  public void shouldHandleRestartWithUnrecoverableForkBlocks_prune(@TempDir final Path tempDir) {
+    testShouldHandleRestartWithUnrecoverableForkBlocks(tempDir, StateStorageMode.PRUNE);
+  }
+
+  private void testShouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart(
+      @TempDir final Path tempDir, final StateStorageMode storageMode) {
+    final long finalizedSlot = 7;
+    final int hotBlockCount = 3;
+    // Setup chains
+    chainBuilder.generateBlocksUpToSlot(finalizedSlot);
+    SignedBlockAndState finalizedBlock = chainBuilder.getBlockAndStateAtSlot(finalizedSlot);
+    final Checkpoint finalizedCheckpoint = getCheckpointForBlock(finalizedBlock.getBlock());
+    final long firstHotBlockSlot =
+        finalizedCheckpoint.getEpochStartSlot(spec).plus(UInt64.ONE).longValue();
+    for (int i = 0; i < hotBlockCount; i++) {
+      chainBuilder.generateBlockAtSlot(firstHotBlockSlot + i);
+    }
+    final long lastSlot = chainBuilder.getLatestSlot().longValue();
+
+    // Setup database
+    createStorage(tempDir.toFile(), storageMode);
+    initGenesis();
+
+    add(chainBuilder.streamBlocksAndStates().collect(Collectors.toSet()));
+
+    // Close database and rebuild from disk
+    restartStorage();
+
+    // Ensure all states are actually regenerated in memory since we expect every state to be stored
+    chainBuilder
+        .streamBlocksAndStates()
+        .forEach(
+            blockAndState -> recentChainData.retrieveBlockState(blockAndState.getRoot()).join());
+
+    justifyAndFinalizeEpoch(finalizedCheckpoint.getEpoch(), finalizedBlock);
+
+    // We should be able to access hot blocks and state
+    final List<SignedBlockAndState> expectedHotBlocksAndStates =
+        chainBuilder.streamBlocksAndStates(finalizedSlot, lastSlot).collect(toList());
+    assertHotBlocksAndStatesInclude(expectedHotBlocksAndStates);
+
+    final Map<Bytes32, BeaconState> historicalStates =
+        chainBuilder
+            .streamBlocksAndStates(0, 6)
+            .collect(Collectors.toMap(SignedBlockAndState::getRoot, SignedBlockAndState::getState));
+
+    switch (storageMode) {
+      case ARCHIVE:
+        assertFinalizedStatesAvailable(historicalStates);
+        break;
+      case PRUNE:
+        assertStatesUnavailable(
+            historicalStates.values().stream().map(BeaconState::getSlot).collect(toList()));
+        break;
+    }
+  }
+
+  private void testShouldHandleRestartWithUnrecoverableForkBlocks(
+      @TempDir final Path tempDir, final StateStorageMode storageMode) {
+    createStorage(tempDir.toFile(), storageMode);
+    final CreateForkChainResult forkChainResult = createForkChain(true);
+
+    // Fork states should be unavailable
+    final UInt64 firstHotBlockSlot = forkChainResult.getFirstHotBlockSlot();
+    final List<Bytes32> unavailableBlockRoots =
+        forkChainResult
+            .getForkChain()
+            .streamBlocksAndStates(4, firstHotBlockSlot.longValue())
+            .map(SignedBlockAndState::getRoot)
+            .collect(Collectors.toList());
+    final List<UInt64> unavailableBlockSlots =
+        forkChainResult
+            .getForkChain()
+            .streamBlocksAndStates(4, firstHotBlockSlot.longValue())
+            .map(SignedBlockAndState::getSlot)
+            .collect(Collectors.toList());
+    assertStatesUnavailable(unavailableBlockSlots);
+    assertBlocksUnavailable(unavailableBlockRoots);
+  }
+
+  private Set<UInt64> getEpochBoundarySlots(final int fromEpoch, final int toEpoch) {
+    final Set<UInt64> epochBoundarySlots = new HashSet<>();
+    for (int i = fromEpoch; i <= toEpoch; i++) {
+      final UInt64 epochSlot = spec.computeStartSlotAtEpoch(UInt64.valueOf(i));
+      epochBoundarySlots.add(epochSlot);
+    }
+    return epochBoundarySlots;
+  }
+
+  private void testShouldPersistOnDisk(final StateStorageMode storageMode) throws Exception {
+    testShouldRecordFinalizedBlocksAndStates(storageMode, false);
   }
 
   public static class CreateForkChainResult {
@@ -989,7 +1572,8 @@ public abstract class AbstractDatabaseTest {
     return new CreateForkChainResult(forkChain, firstHotBlockSlot);
   }
 
-  public void testStartupFromNonGenesisState(final StateStorageMode storageMode) {
+  public void testStartupFromNonGenesisState(final StateStorageMode storageMode)
+      throws IOException {
     createStorage(storageMode);
 
     // Set up database from an anchor point
@@ -1017,17 +1601,17 @@ public abstract class AbstractDatabaseTest {
   }
 
   @Test
-  public void startupFromNonGenesisStateAndFinalizeNewCheckpoint_prune() {
+  public void startupFromNonGenesisStateAndFinalizeNewCheckpoint_prune() throws IOException {
     testStartupFromNonGenesisStateAndFinalizeNewCheckpoint(StateStorageMode.PRUNE);
   }
 
   @Test
-  public void startupFromNonGenesisStateAndFinalizeNewCheckpoint_archive() {
+  public void startupFromNonGenesisStateAndFinalizeNewCheckpoint_archive() throws IOException {
     testStartupFromNonGenesisStateAndFinalizeNewCheckpoint(StateStorageMode.ARCHIVE);
   }
 
   @Test
-  void shouldStoreAndRetrieveVotes() {
+  void shouldStoreAndRetrieveVotes() throws IOException {
     final DataStructureUtil dataStructureUtil = new DataStructureUtil(spec);
     createStorage(StateStorageMode.PRUNE);
     assertThat(database.getVotes()).isEmpty();
@@ -1053,7 +1637,7 @@ public abstract class AbstractDatabaseTest {
   }
 
   public void testStartupFromNonGenesisStateAndFinalizeNewCheckpoint(
-      final StateStorageMode storageMode) {
+      final StateStorageMode storageMode) throws IOException {
     createStorage(storageMode);
 
     // Set up database from an anchor point
@@ -1094,15 +1678,8 @@ public abstract class AbstractDatabaseTest {
     return stateRoot;
   }
 
-  public void testShouldRecordFinalizedBlocksAndStates(
-      final StateStorageMode storageMode, final boolean batchUpdate) {
-    testShouldRecordFinalizedBlocksAndStates(storageMode, batchUpdate, this::createStorage);
-  }
-
   protected void testShouldRecordFinalizedBlocksAndStates(
-      final StateStorageMode storageMode,
-      final boolean batchUpdate,
-      Consumer<StateStorageMode> initializeDatabase) {
+      final StateStorageMode storageMode, final boolean batchUpdate) throws IOException {
     // Setup chains
     // Both chains share block up to slot 3
     final ChainBuilder primaryChain = ChainBuilder.create(spec, VALIDATOR_KEYS);
@@ -1122,7 +1699,7 @@ public abstract class AbstractDatabaseTest {
     forkChain.generateBlockAtSlot(hotSlot);
 
     // Setup database
-    initializeDatabase.accept(storageMode);
+    createStorage(storageMode);
     initGenesis();
 
     final Set<SignedBlockAndState> allBlocksAndStates =
