--- storage/src/test/java/tech/pegasys/teku/storage/server/kvstore/AbstractKvStoreDatabaseTest.java
@@ -1,360 +1,0 @@
-/*
- * Copyright ConsenSys Software Inc., 2022
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.storage.server.kvstore;
-
-import static java.util.stream.Collectors.toList;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.mock;
-import static tech.pegasys.teku.infrastructure.async.SafeFutureAssert.assertThatSafeFuture;
-
-import java.nio.file.Path;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import org.apache.tuweni.bytes.Bytes32;
-import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
-import tech.pegasys.teku.dataproviders.lookup.BlockProvider;
-import tech.pegasys.teku.dataproviders.lookup.StateAndBlockSummaryProvider;
-import tech.pegasys.teku.ethereum.pow.api.MinGenesisTimeBlockEvent;
-import tech.pegasys.teku.infrastructure.async.AsyncRunner;
-import tech.pegasys.teku.infrastructure.async.SafeFuture;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.spec.TestSpecFactory;
-import tech.pegasys.teku.spec.datastructures.blocks.BlockAndCheckpointEpochs;
-import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.spec.datastructures.blocks.SignedBlockAndState;
-import tech.pegasys.teku.spec.datastructures.state.Checkpoint;
-import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
-import tech.pegasys.teku.spec.util.DataStructureUtil;
-import tech.pegasys.teku.storage.api.OnDiskStoreData;
-import tech.pegasys.teku.storage.server.AbstractStorageBackedDatabaseTest;
-import tech.pegasys.teku.storage.server.ShuttingDownException;
-import tech.pegasys.teku.storage.server.StateStorageMode;
-import tech.pegasys.teku.storage.server.kvstore.dataaccess.KvStoreFinalizedDao;
-import tech.pegasys.teku.storage.server.kvstore.dataaccess.KvStoreFinalizedDao.FinalizedUpdater;
-import tech.pegasys.teku.storage.server.kvstore.dataaccess.KvStoreHotDao;
-import tech.pegasys.teku.storage.server.kvstore.dataaccess.KvStoreHotDao.HotUpdater;
-import tech.pegasys.teku.storage.store.StoreBuilder;
-import tech.pegasys.teku.storage.store.UpdatableStore;
-import tech.pegasys.teku.storage.store.UpdatableStore.StoreTransaction;
-
-public abstract class AbstractKvStoreDatabaseTest extends AbstractStorageBackedDatabaseTest {
-
-  @Test
-  public void shouldThrowIfClosedDatabaseIsModified_setGenesis() throws Exception {
-    database.close();
-    assertThatThrownBy(() -> database.storeInitialAnchor(genesisAnchor))
-        .isInstanceOf(ShuttingDownException.class);
-  }
-
-  @Test
-  public void shouldThrowIfClosedDatabaseIsModified_update() throws Exception {
-    database.storeInitialAnchor(genesisAnchor);
-    database.close();
-
-    final SignedBlockAndState newValue = chainBuilder.generateBlockAtSlot(1);
-    // Sanity check
-    assertThatSafeFuture(store.retrieveBlockState(newValue.getRoot()))
-        .isCompletedWithEmptyOptional();
-    final StoreTransaction transaction = recentChainData.startStoreTransaction();
-    transaction.putBlockAndState(newValue);
-
-    final SafeFuture<Void> result = transaction.commit();
-    assertThatThrownBy(result::get).hasCauseInstanceOf(ShuttingDownException.class);
-  }
-
-  @Test
-  public void createMemoryStore_priorToGenesisTime() {
-    database.storeInitialAnchor(genesisAnchor);
-
-    final Optional<OnDiskStoreData> maybeData =
-        ((KvStoreDatabase) database).createMemoryStore(() -> 0L);
-    assertThat(maybeData).isNotEmpty();
-
-    final OnDiskStoreData data = maybeData.get();
-    final UpdatableStore store =
-        StoreBuilder.create()
-            .metricsSystem(new NoOpMetricsSystem())
-            .specProvider(spec)
-            .time(data.getTime())
-            .anchor(data.getAnchor())
-            .genesisTime(data.getGenesisTime())
-            .latestFinalized(data.getLatestFinalized())
-            .finalizedOptimisticTransitionPayload(data.getFinalizedOptimisticTransitionPayload())
-            .justifiedCheckpoint(data.getJustifiedCheckpoint())
-            .bestJustifiedCheckpoint(data.getBestJustifiedCheckpoint())
-            .blockInformation(data.getBlockInformation())
-            .votes(data.getVotes())
-            .asyncRunner(mock(AsyncRunner.class))
-            .blockProvider(mock(BlockProvider.class))
-            .stateProvider(mock(StateAndBlockSummaryProvider.class))
-            .build();
-
-    assertThat(store.getTimeSeconds()).isEqualTo(genesisTime);
-  }
-
-  @Test
-  public void shouldThrowIfClosedDatabaseIsRead_createMemoryStore() throws Exception {
-    database.storeInitialAnchor(genesisAnchor);
-    database.close();
-
-    assertThatThrownBy(database::createMemoryStore).isInstanceOf(ShuttingDownException.class);
-  }
-
-  @Test
-  public void shouldThrowIfClosedDatabaseIsRead_getSlotForFinalizedBlockRoot() throws Exception {
-    database.storeInitialAnchor(genesisAnchor);
-    database.close();
-
-    assertThatThrownBy(() -> database.getSlotForFinalizedBlockRoot(Bytes32.ZERO))
-        .isInstanceOf(ShuttingDownException.class);
-  }
-
-  @Test
-  public void shouldThrowIfClosedDatabaseIsRead_getSignedBlock() throws Exception {
-    database.storeInitialAnchor(genesisAnchor);
-    database.close();
-
-    assertThatThrownBy(() -> database.getSignedBlock(genesisCheckpoint.getRoot()))
-        .isInstanceOf(ShuttingDownException.class);
-  }
-
-  @Test
-  public void shouldThrowIfClosedDatabaseIsRead_streamFinalizedBlocks() throws Exception {
-    database.storeInitialAnchor(genesisAnchor);
-    database.close();
-
-    assertThatThrownBy(() -> database.streamFinalizedBlocks(UInt64.ZERO, UInt64.ONE))
-        .isInstanceOf(ShuttingDownException.class);
-  }
-
-  @Test
-  public void shouldThrowIfClosedDatabaseIsRead_streamFinalizedBlocksShuttingDown()
-      throws Exception {
-    database.storeInitialAnchor(genesisAnchor);
-    try (final Stream<SignedBeaconBlock> stream =
-        database.streamFinalizedBlocks(UInt64.ZERO, UInt64.valueOf(1000L))) {
-      database.close();
-      assertThatThrownBy(stream::findAny).isInstanceOf(ShuttingDownException.class);
-    }
-  }
-
-  @Test
-  public void shouldThrowIfTransactionModifiedAfterDatabaseIsClosed_updateHotDao()
-      throws Exception {
-    database.storeInitialAnchor(genesisAnchor);
-
-    try (final KvStoreHotDao.HotUpdater updater = hotUpdater()) {
-      SignedBlockAndState newBlock = chainBuilder.generateNextBlock();
-      database.close();
-      assertThatThrownBy(
-              () -> updater.addHotBlock(BlockAndCheckpointEpochs.fromBlockAndState(newBlock)))
-          .isInstanceOf(ShuttingDownException.class);
-    }
-  }
-
-  private HotUpdater hotUpdater() {
-    return ((KvStoreDatabase) database).dao.hotUpdater();
-  }
-
-  @Test
-  public void shouldThrowIfTransactionModifiedAfterDatabaseIsClosed_updateFinalizedDao()
-      throws Exception {
-    database.storeInitialAnchor(genesisAnchor);
-
-    try (final KvStoreFinalizedDao.FinalizedUpdater updater = finalizedUpdater()) {
-      SignedBlockAndState newBlock = chainBuilder.generateNextBlock();
-      database.close();
-      assertThatThrownBy(() -> updater.addFinalizedBlock(newBlock.getBlock()))
-          .isInstanceOf(ShuttingDownException.class);
-    }
-  }
-
-  private FinalizedUpdater finalizedUpdater() {
-    return ((KvStoreDatabase) database).dao.finalizedUpdater();
-  }
-
-  @Test
-  public void shouldThrowIfTransactionModifiedAfterDatabaseIsClosed_updateEth1Dao()
-      throws Exception {
-    database.storeInitialAnchor(genesisAnchor);
-
-    final DataStructureUtil dataStructureUtil =
-        new DataStructureUtil(TestSpecFactory.createDefault());
-    try (final HotUpdater updater = hotUpdater()) {
-      final MinGenesisTimeBlockEvent genesisTimeBlockEvent =
-          dataStructureUtil.randomMinGenesisTimeBlockEvent(1);
-      database.close();
-      assertThatThrownBy(() -> updater.addMinGenesisTimeBlock(genesisTimeBlockEvent))
-          .isInstanceOf(ShuttingDownException.class);
-    }
-  }
-
-  @Test
-  public void shouldThrowIfClosedDatabaseIsRead_getHistoricalState() throws Exception {
-    // Store genesis
-    database.storeInitialAnchor(genesisAnchor);
-    // Add a new finalized block to supersede genesis
-    final SignedBlockAndState newBlock = chainBuilder.generateBlockAtSlot(1);
-    final Checkpoint newCheckpoint = getCheckpointForBlock(newBlock.getBlock());
-    final StoreTransaction transaction = recentChainData.startStoreTransaction();
-    transaction.putBlockAndState(newBlock);
-    transaction.setFinalizedCheckpoint(newCheckpoint, false);
-    transaction.commit().ifExceptionGetsHereRaiseABug();
-    // Close db
-    database.close();
-
-    assertThatThrownBy(
-            () ->
-                database.getLatestAvailableFinalizedState(
-                    genesisCheckpoint.getEpochStartSlot(spec)))
-        .isInstanceOf(ShuttingDownException.class);
-  }
-
-  @Test
-  public void shouldThrowIfTransactionModifiedAfterDatabaseIsClosedFromAnotherThread()
-      throws Exception {
-
-    for (int i = 0; i < 20; i++) {
-      createStorage(StateStorageMode.PRUNE);
-      database.storeInitialAnchor(genesisAnchor);
-
-      try (final KvStoreHotDao.HotUpdater updater = hotUpdater()) {
-        SignedBlockAndState newBlock = chainBuilder.generateNextBlock();
-
-        final Thread dbCloserThread =
-            new Thread(
-                () -> {
-                  try {
-                    database.close();
-                  } catch (Exception e) {
-                    throw new RuntimeException(e);
-                  }
-                });
-
-        dbCloserThread.start();
-        try {
-          updater.addHotBlock(BlockAndCheckpointEpochs.fromBlockAndState(newBlock));
-        } catch (Exception e) {
-          assertThat(e).isInstanceOf(ShuttingDownException.class);
-        }
-
-        dbCloserThread.join(500);
-      }
-    }
-  }
-
-  @Test
-  public void shouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart__archive(
-      @TempDir final Path tempDir) {
-    testShouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart(tempDir, StateStorageMode.ARCHIVE);
-  }
-
-  @Test
-  public void shouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart__prune(
-      @TempDir final Path tempDir) {
-    testShouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart(tempDir, StateStorageMode.PRUNE);
-  }
-
-  private void testShouldPruneHotBlocksOlderThanFinalizedSlotAfterRestart(
-      @TempDir final Path tempDir, final StateStorageMode storageMode) {
-    final long finalizedSlot = 7;
-    final int hotBlockCount = 3;
-    // Setup chains
-    chainBuilder.generateBlocksUpToSlot(finalizedSlot);
-    SignedBlockAndState finalizedBlock = chainBuilder.getBlockAndStateAtSlot(finalizedSlot);
-    final Checkpoint finalizedCheckpoint = getCheckpointForBlock(finalizedBlock.getBlock());
-    final long firstHotBlockSlot =
-        finalizedCheckpoint.getEpochStartSlot(spec).plus(UInt64.ONE).longValue();
-    for (int i = 0; i < hotBlockCount; i++) {
-      chainBuilder.generateBlockAtSlot(firstHotBlockSlot + i);
-    }
-    final long lastSlot = chainBuilder.getLatestSlot().longValue();
-
-    // Setup database
-    createStorage(tempDir.toFile(), storageMode);
-    initGenesis();
-
-    add(chainBuilder.streamBlocksAndStates().collect(Collectors.toSet()));
-
-    // Close database and rebuild from disk
-    restartStorage();
-
-    // Ensure all states are actually regenerated in memory since we expect every state to be stored
-    chainBuilder
-        .streamBlocksAndStates()
-        .forEach(
-            blockAndState -> recentChainData.retrieveBlockState(blockAndState.getRoot()).join());
-
-    justifyAndFinalizeEpoch(finalizedCheckpoint.getEpoch(), finalizedBlock);
-
-    // We should be able to access hot blocks and state
-    final List<SignedBlockAndState> expectedHotBlocksAndStates =
-        chainBuilder.streamBlocksAndStates(finalizedSlot, lastSlot).collect(toList());
-    assertHotBlocksAndStatesInclude(expectedHotBlocksAndStates);
-
-    final Map<Bytes32, BeaconState> historicalStates =
-        chainBuilder
-            .streamBlocksAndStates(0, 6)
-            .collect(Collectors.toMap(SignedBlockAndState::getRoot, SignedBlockAndState::getState));
-
-    switch (storageMode) {
-      case ARCHIVE:
-        assertFinalizedStatesAvailable(historicalStates);
-        break;
-      case PRUNE:
-        assertStatesUnavailable(
-            historicalStates.values().stream().map(BeaconState::getSlot).collect(toList()));
-        break;
-    }
-  }
-
-  @Test
-  public void shouldHandleRestartWithUnrecoverableForkBlocks_archive(@TempDir final Path tempDir) {
-    testShouldHandleRestartWithUnrecoverableForkBlocks(tempDir, StateStorageMode.ARCHIVE);
-  }
-
-  @Test
-  public void shouldHandleRestartWithUnrecoverableForkBlocks_prune(@TempDir final Path tempDir) {
-    testShouldHandleRestartWithUnrecoverableForkBlocks(tempDir, StateStorageMode.PRUNE);
-  }
-
-  private void testShouldHandleRestartWithUnrecoverableForkBlocks(
-      @TempDir final Path tempDir, final StateStorageMode storageMode) {
-    createStorage(tempDir.toFile(), storageMode);
-    final CreateForkChainResult forkChainResult = createForkChain(true);
-
-    // Fork states should be unavailable
-    final UInt64 firstHotBlockSlot = forkChainResult.getFirstHotBlockSlot();
-    final List<Bytes32> unavailableBlockRoots =
-        forkChainResult
-            .getForkChain()
-            .streamBlocksAndStates(4, firstHotBlockSlot.longValue())
-            .map(SignedBlockAndState::getRoot)
-            .collect(Collectors.toList());
-    final List<UInt64> unavailableBlockSlots =
-        forkChainResult
-            .getForkChain()
-            .streamBlocksAndStates(4, firstHotBlockSlot.longValue())
-            .map(SignedBlockAndState::getSlot)
-            .collect(Collectors.toList());
-    assertStatesUnavailable(unavailableBlockSlots);
-    assertBlocksUnavailable(unavailableBlockRoots);
-  }
-}
