+++ storage/src/test/java/tech/pegasys/teku/storage/server/leveldb/TrieLevelDbDatabaseTest.java
@@ -20,13 +20,13 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import tech.pegasys.teku.storage.server.DatabaseVersion;
 import tech.pegasys.teku.storage.server.StateStorageMode;
-import tech.pegasys.teku.storage.server.kvstore.AbstractKvStoreDatabaseWithHotStatesTest;
+import tech.pegasys.teku.storage.server.kvstore.AbstractDatabaseTest;
 import tech.pegasys.teku.storage.storageSystem.FileBackedStorageSystemBuilder;
 import tech.pegasys.teku.storage.storageSystem.StorageSystem;
 import tech.pegasys.teku.storage.store.StoreConfig;
 
 @Disabled("https://github.com/ConsenSys/teku/issues/4173 still in progress")
-public class TrieLevelDbDatabaseTest extends AbstractKvStoreDatabaseWithHotStatesTest {
+public class TrieLevelDbDatabaseTest extends AbstractDatabaseTest {
 
   @BeforeAll
   static void setUp() {
