+++ storage/src/test/java/tech/pegasys/teku/storage/server/rocksdb/V4RocksDbDatabaseTest.java
@@ -19,12 +19,12 @@ import java.io.File;
 import org.junit.jupiter.api.BeforeAll;
 import tech.pegasys.teku.storage.server.DatabaseVersion;
 import tech.pegasys.teku.storage.server.StateStorageMode;
-import tech.pegasys.teku.storage.server.kvstore.AbstractKvStoreDatabaseWithHotStatesTest;
+import tech.pegasys.teku.storage.server.kvstore.AbstractDatabaseTest;
 import tech.pegasys.teku.storage.storageSystem.FileBackedStorageSystemBuilder;
 import tech.pegasys.teku.storage.storageSystem.StorageSystem;
 import tech.pegasys.teku.storage.store.StoreConfig;
 
-public class V4RocksDbDatabaseTest extends AbstractKvStoreDatabaseWithHotStatesTest {
+public class V4RocksDbDatabaseTest extends AbstractDatabaseTest {
 
   @BeforeAll
   static void setUp() {
